Konstructor : Enable downloaded archive to be reseted

+++ src/konstructor.py
@@ -702,7 +702,7 @@ You can try the following:
                 os.rmdir(files[0])
 
     @staticmethod
-    def download(location, downloadDir=None, destinationDir=None):
+    def download(location, downloadDir=None, destinationDir=None, useExistingArchive=False):
         import types
 
         if downloadDir is None:
@@ -718,23 +718,24 @@ You can try the following:
             return location.download(destinationDir)
         else:
             if location.startswith("http"):
-                return Utils._httpDownload(location, downloadDir, destinationDir)
+                return Utils._httpDownload(location, downloadDir, destinationDir, useExistingArchive=useExistingArchive)
             else:
                 Utils.exit("Protocol not supported for downloading " + location)
 
     @staticmethod
-    def _httpDownload(url, downloadDir, destinationDir):
+    def _httpDownload(url, downloadDir, destinationDir, useExistingArchive=False):
         import urllib2
         import tempfile
 
         file_name = url.split('/')[-1]
         extractDir = os.path.join(downloadDir, file_name)
         exists = os.path.exists(file_name)
-        if exists:
+        if exists and not useExistingArchive:
             if Utils.promptYesNo("The downloadfile %s is already present, download a fresh version ?" % (file_name)):
                 Log.debug("Removing %s" % (file_name))
                 os.unlink(file_name)
                 exists = False
+
         if not exists:
             u = urllib2.urlopen(url)
             f = open(extractDir, "wb")
@@ -751,6 +752,7 @@ You can try the following:
                 file_size_dl += len(buff)
                 f.write(buff)
             f.close()
+
         if destinationDir:
             Log.info("Extracting %s" % (file_name))
             Utils.extract(os.path.join(downloadDir, file_name), destinationDir)
@@ -935,7 +937,7 @@ class Dep:
                 Log.debug("Build discarded because of --ignore-build flag")
                 self.needBuild = False
 
-    def download(self):
+    def download(self, useExistingArchive=False):
         import shutil
         if not self.needDownload:
             return
@@ -957,7 +959,7 @@ class Dep:
             shutil.rmtree(self.extractDir)
 
         Log.info("Downloading %s" % (self.name))
-        Utils.download(self.options["location"], downloadDir=".", destinationDir=self.extractDir)
+        Utils.download(self.options["location"], downloadDir=".", destinationDir=self.extractDir, useExistingArchive=useExistingArchive)
 
         if self.linkDir:
             # Make the dep directory point to the directory matching the configuration
@@ -965,24 +967,45 @@ class Dep:
 
         self.cache.setConfig(self.name + "-download", self.downloadConfig);
 
+    def revertLocalChanges(self):
+        isDepString = False
+        isDepObject = False
+
+        if type(self.options["location"]) == str:
+            isDepString = True
+        elif hasattr(self.options["location"], "reset"):
+            isDepObject =True
+
+        canRevert = isDepObject or isDepString
+
+        if not canRevert:
+            Utils.exit("Failed to apply patch and dependency cannot be reverted.")
+
+        if Utils.promptYesNo("Revert local changes and try to apply patch again ?"):
+            if isDepObject:
+                Log.info("is dep object")
+                with Utils.Chdir(self.extractDir):
+                    self.options["location"].reset()
+            else:
+                Log.info("calling download again")
+                self.needDownload = True
+                self.download(useExistingArchive=True)
+            return self.patch(noReset=True)
+        else:
+            Log.info("Not reverting, patch cannot be applied")
+            return False
+
     def patch(self, noReset=False):
         if "patchs" not in self.options:
-            return
+            return True
 
         Log.debug("Patching " + self.name)
         for p in self.options["patchs"]:
             if not Utils.patch(self.name, p):
-                Log.error("Failed to apply patch")
-                if hasattr(self.options["location"], "reset") and not noReset:
-                    if Utils.promptYesNo("Revert local changes and try to apply patch again ?"):
-                        with Utils.Chdir(self.extractDir):
-                            self.options["location"].reset()
-                        self.patch(noReset=True)
-                        return
-                    else:
-                        Utils.exit("Failed to patch")
-                else:
-                    Utils.exit("Failed to patch")
+                Log.error("Failed to apply patch. Trying to revert local changes")
+                if not noReset and not self.revertLocalChanges():
+                    Utils.exit("Failed to apply patch")
+        return True
 
     def _getDir(self):
         newDir = "."
